+++ src/main/java/com/finances/api/domain/usecases/createTransaction/CreateTransactionUseCaseDto.java
@@ -9,7 +9,7 @@ import java.util.UUID;
 public record CreateTransactionUseCaseDto(
         @NotNull(message = "Transaction value cannot be null")
         @Positive(message = "Transaction value must be greater than 0")
-        Double value,
+        float value,
 
         @NotNull(message = "Transaction type cannot be null")
         @Size(min = 1, max = 50, message = "Transaction type must be between 1 and 50 characters")
