+++ src/main/java/com/finances/api/services/protocols/createTransactionService/CreateTransactionServiceDto.java
@@ -3,7 +3,7 @@ package com.finances.api.services.protocols.createTransactionService;
 import java.util.UUID;
 
 public record CreateTransactionServiceDto(
-        Double value,
+        float value,
         String type,
         UUID fromAccountId,
         UUID targetAccountId
